corrected notifications time

+++ chrome/android/java/src/org/chromium/chrome/browser/toolbar/ToolbarTablet.java
@@ -672,7 +672,7 @@ public class ToolbarTablet
     public void OnGetCurrentBalanceReport(String[] report) {}
 
     @Override
-    public void OnNotificationAdded(String id, int type, int timestamp,
+    public void OnNotificationAdded(String id, int type, long timestamp,
             String[] args) {
         if (mBraveRewardsNativeWorker != null) {
             mBraveRewardsNativeWorker.GetAllNotifications();
@@ -701,7 +701,7 @@ public class ToolbarTablet
     }
 
     @Override
-    public void OnGetLatestNotification(String id, int type, int timestamp,
+    public void OnGetLatestNotification(String id, int type, long timestamp,
             String[] args) {}
 
     @Override
